begin porting to CMake using stm32-base
