+++ README.md
@@ -1,12 +1,3 @@
-Description
------------
+# stm32-enc28j60
 
-An implementation of ENC28J60 driver for STM32 MCU. Tested only on sending UDP packets and only on MCUs STM32F091xC, STM32F030x6 and STM32F103xB. If you need to use another STM32 MCU then don't forget to edit file `inc/stm32includes.h`.
-
-Working examples
-----------------
-
-* [dc-thermal-logger](https://devel.mephi.ru/dyokunev/dc-thermal-logger/src/master/collector/firmware/Src/main.c)
-* [CardLogger](https://devel.mephi.ru/AMGladtsin/STM32_Devel/src/master/SOFT/CardLogger/Src/main.c)
-* [enc28j60-helloWorld](https://devel.mephi.ru/dyokunev/nucleo-f091rc-rs232-enc28j60/src/master/Src/main.c)
-
+STM32 ENC28J60 ethernet driver, ported for CMake
