+++ src/dnslkup.h
