+++ stm32-env-example.ps1
@@ -1,0 +1,3 @@
+$env:STM32_TOOLCHAIN_PATH = "<your path to stm32-cmake>\cmake\stm32_gcc.cmake"
+$env:STM32_TARGET_TRIPLET = "arm-none-eabi"
+$env:STM32_CUBE_F4_PATH="<path to stm32 cube firmware drives>"
